Update overview.mdx
For the code example to work, the tool must either have an inputSchema or an outputSchema, not just a Schema. 

Also, it seems like Id is required.

For clarity, let's add a Context wrapper around your metadata. It will help avoid issues when additional metadata elements are provided and help better destructuring syntax